+++ adminer/include/html.inc.php
@@ -150,7 +150,7 @@ function js_escape(string $string): string {
 function pagination(int $page, ?int $current): string {
 	return " " . ($page == $current
 		? $page + 1
-		: '<a href="' . h(remove_from_uri("page") . ($page ? "&page=$page" . ($_GET["next"] ? "&next=" . urlencode($_GET["next"]) : "") : "")) . '">' . ($page + 1) . "</a>"
+		: '<a href="' . h(remove_from_uri("page|next") . ($page ? "&page=$page" . ($_GET["next"] ? "&next=" . urlencode($_GET["next"]) : "") : "")) . '">' . ($page + 1) . "</a>"
 	);
 }